+++ packages/flame/lib/src/components/base_component.dart
@@ -27,6 +27,7 @@ abstract class BaseComponent extends Component {
   /// If the component has a parent it will be set here
   BaseComponent? _parent;
 
+  @override
   BaseComponent? get parent => _parent;
 
   /// The children list shouldn't be modified directly, that is why an
@@ -56,6 +57,8 @@ abstract class BaseComponent extends Component {
           fontSize: 12,
         ),
       );
+
+  BaseComponent({int priority = 0}) : super(priority: priority);
 
   /// This method is called periodically by the game engine to request that your component updates itself.
   ///
@@ -171,6 +174,17 @@ abstract class BaseComponent extends Component {
     }
   }
 
+  Future<void> addChildren(
+    Iterable<Component> children, {
+    Game? gameRef,
+  }) async {
+    await Future.wait(
+      children.map(
+        (child) => addChild(child, gameRef: gameRef),
+      ),
+    );
+  }
+
   bool removeChild(Component c) {
     return _children.remove(c);
   }
@@ -180,6 +194,8 @@ abstract class BaseComponent extends Component {
   }
 
   bool containsChild(Component c) => _children.contains(c);
+
+  void reorderChildren() => _children.rebalanceAll();
 
   /// This method first calls the passed handler on the leaves in the tree,
   /// the children without any children of their own.